+++ backend/src/routes/index.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const controller = require('../controllers');
-
-const router = express.Router();
-
-router.get('/api/health', controller.health);
-router.get('/api/get-temp/:city', controller.getTemp);
-
-module.exports = router;